Rename modules and files to have consistent casing
