+++ app/messages.py
@@ -1,6 +1,5 @@
 from enum import Enum
 from typing import List
-
 from pydantic import BaseModel
