+++ models/human_detection/README.md
@@ -1,10 +1,10 @@
-# HumanDetecton
+# human_detection
 
 ## Training
 
 To train run the following command, using the same venv as DMAS
 ```
-python3 HumanDetector.py
+python3 human_detector.py
 ```
 
 This will save the weights in the weights directory
